rename MovieCharactersContext to CharactersContext
